+++ src/apis/extractPages.js
@@ -1,7 +1,7 @@
 /**
  * Extract pages from the current document
  * @method WebViewer#extractPages
- * @param {Array<number>} pageIndexToExtract An array of pages to extract
+ * @param {Array<number>} pageIndexToExtract An array of pages to extract from the document. Annotations on the pages are included
  * @return {object} A promise that resolve to a <a href='https://developer.mozilla.org/en-US/docs/Web/API/File' target='_blank'>File object</a>
  * @example // 6.0 and after
 WebViewer(...)
