+++ src/components/DocumentControls/DocumentControls.scss
@@ -42,6 +42,7 @@
         .Button{
             display: inline-flex;
             height: 100%;
+            padding: 0px;
             min-width: $icon-width;
             flex-grow: 1;
