+++ src/helpers/getPageArrayFromString.js
